refactoring small issues in ArticleParamsForm and App
