+++ src/components/app/App.tsx
@@ -1,0 +1,25 @@
+import { CSSProperties, useState } from 'react';
+import { defaultArticleState } from 'src/constants/articleProps';
+import styles from './App.module.scss';
+import { ArticleParamsForm } from '../article-params-form/ArticleParamsForm';
+import { Article } from '../article/Article';
+
+export const App = () => {
+	const [articleState, setArticleState] = useState(defaultArticleState);
+	return (
+		<main
+			className={styles.main}
+			style={
+				{
+					'--font-family': articleState.fontFamilyOption.value,
+					'--font-size': articleState.fontSizeOption.value,
+					'--font-color': articleState.fontColor.value,
+					'--container-width': articleState.contentWidth.value,
+					'--bg-color': articleState.backgroundColor.value,
+				} as CSSProperties
+			}>
+			<ArticleParamsForm setArticleState={setArticleState} />
+			<Article />
+		</main>
+	);
+};
