+++ src/index.tsx
@@ -1,39 +1,12 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
-
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import { defaultArticleState } from './constants/articleProps';
+import { StrictMode } from 'react';
 
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
+
+import { App } from './components/app/App';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-export const App = () => {
-	const [articleState, setArticleState] = useState(defaultArticleState);
-	return (
-		<main
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': articleState.fontFamilyOption.value,
-					'--font-size': articleState.fontSizeOption.value,
-					'--font-color': articleState.fontColor.value,
-					'--container-width': articleState.contentWidth.value,
-					'--bg-color': articleState.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm
-				articleState={articleState}
-				setArticleState={setArticleState}
-			/>
-			<Article />
-		</main>
-	);
-};
 
 root.render(
 	<StrictMode>
